pyproject.toml
[tool.poetry]
name = "tappet"
version = "0.1.0"
description = "TUI HTTP Client with request management"
authors = ["nattoujam <public.kyuuanago@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
textual = "^0.47.0"
click = "^8.1.7"
httpx = "^0.27.0"
pyyaml = "^6.0.1"
pyperclip = "^1.8.2"

[tool.poetry.scripts]
tappet = "tappet.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "autopep8 (>=2.3.2,<3.0.0)",
    "isort (<6)"
]
